Cleaning up.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@2132 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ Testprog/stereoIIDC/createFeatureCmds.cc
@@ -40,7 +40,7 @@ struct Feature
 {
     IIDCCamera::Feature	feature;
     const char*		name;
-    int			prop[3];
+    float		prop[3];
 };
 static Feature		feature[] =
 {
